make `Executables` class
